Extract PRAttributionRecorder from IssueTracker (CROW-1094)
Move PR→session attribution and the #694 rework/revert signals
(recordPRAttribution, updatePRAttributions, captureChangedFilesForNewMerges,
scanDefaultBranchesForReverts, detectPostMergeFixes + the pure decision statics
and constants) into their own @MainActor collaborator. It owns lastRevertScanAt
and changedFilesFetchAttempted; persistence still routes through the shared,
injected JSONStore (ADR 0012 / #728) via an unowned back-reference to the
tracker. refresh() and prHasCrowAuthoredCommit call it in the same order.

Behavior-preserving structural split — no test assertion changes.
swift test --package-path Packages/CrowEngine: 789 tests pass.


Crow-Session: 18A4CD55-8B33-48E4-B5EB-E710314DF5E5
